First draft models
Following a workshop with the team I've revised the models and reformatted them to reflect the questions they ask. They may need to be changed in the future to better reflect reality, but this is a good first step.

This commit also includes a change to the docker-compose, since I generated a bunch of migrations that were lost since I wasn't mapping the folder. Sigh.