Add functionality to report on Boolean attributes
I've learned about getattr(object, attr: str) today and it's very cool. I've also written a subclass to deal with the challenge of reports for the boolean attributes in the ProtectedCharacteristic group. I could equally have made two new tables and used the existing report, but I think this way is neater for now. This new report should work for caring responsibilities and disabilities