Add neat backref
This extends the mixin so I don't need to keep typing the same boilerplate relationship code. Now anything with this mixin automatically has a relationship 'candidates' that gives a list of candidates associated to that Model's instance. In the opposite direction, each Candidate now has a relationship that is automatically generated from the table name. It's literally the coolest thing ever