Add methods for filtering out non-eligible candidates
This is reasonably neat. There's now a method on the parent class, PromotionReport, that looks out the eligible candidates. They're all the people in that intake year who are signed up to the requested programme. This is then used in the subclasses to produce the correct numbers for the reports and reduce boilerplate code.